+++ main_funcs.py
@@ -1,45 +1,32 @@
 import sys
-import movement as mo
+import os
+
+import menus as me
 import inventory as i
 import weapons as w
 import your_character as yc
 
 
-# menus are important. Let's add them.
-menus ={
-    "main": ["walk", "your character", "inventory"],
-    "walk": ["north", "south", "east", "west"],
-    "yes_no": ["yes", "no"],
-    "your character": ["info", "set name", "set hero name"],
-    "inventory": [],
-} 
-
-
-# Making said menus functional.
-# Function to display menus. 
-def display_menu(name):
-    """Displays chosen menu to user."""
-    menuvalues = menus[name]
-    if name == "main":
-        print("What would you like to do? ")
-    elif name == "walk":
-        print("Which direction would you like to go?")
-    elif name == "your character":
-        print("Choose an action to take relating to your character.")
-    elif name == "yes_no":
-        print("Which option would you like to select?")
-    [print(action.title()) for action in menuvalues]
+# function to save when user quits
+def user_quitting():
+    """
+    Prints a message when the user quits, saves, then exits the program.
+    """
+    print("Ok, farewell, future X-Man.")
+    from saving import save
+    save()
+    sys.exit()
 
 
 # User input function with error handling.
 def user_action(section):
     """Gets user input and checks for bogus answers."""
-    menuvalues = menus[section]
+    menuvalues = me.menus[section]
     correction = "none"
     action_inp = input("Enter any of these, or \"quit\" to exit: ")
     if action_inp in menuvalues or action_inp == "quit":
         if action_inp == "quit":
-            print("Ok, farewell, future X-Man.")
+            user_quitting()
             sys.exit()
         else:
             return action_inp
@@ -53,14 +40,13 @@ def user_action(section):
                 print("Sorry, this option is also incorrect. Please try again.")
             else:
                 print("Ok, you must be joking. Please enter a VALID option.")
-            display_menu(section)
+            me.display_menu(section)
             attempts += 1
             action_inp = input("Enter any of these, or \"quit\" to exit: ")
             if action_inp in menuvalues or action_inp == "quit":
                 correct = 'yes'
                 if action_inp == "quit":
-                    print("Ok, farewell, future X-Man.")
-                    sys.exit()
+                    user_quitting()
                 else:
                     return action_inp
     
@@ -68,15 +54,42 @@ def user_action(section):
 # Function which sends user back to main menu.
 def main_menu():
     """For within other functions. Goes back to main menu."""
-    display_menu("main")
-    m = user_action("main")
-    return m
+    me.display_menu("main")
+    m_act = user_action("main")
+    return m_act
+
+
+# function to save thanks to stupid circular imports.
+def save_game():
+    from saving import save
+    save()
+
+
+# Function to load a save.
+def check_save():
+    """
+    Asks user if they would like to start a game and acts accordingly.
+    """
+    from saving import savefile_path
+    if os.path.exists(savefile_path) == True:
+        print("It looks like you have an existing game. \n Would you like to continue?")
+        me.display_menu("yes_no")
+        choice = user_action("yes_no")
+        if choice == "yes":
+            from saving import load_save
+            load_save()
+        else:
+            print("Ok, starting over.")
+            yc.create_char()
+    else:
+        yc.create_char()
 
 
 # menu interactions 
 def menu_actions(menu):
     """Enables menu interaction."""
     if menu == "walk":
+        import movement as mo
         mo.move(mo.row, mo.col)
         mo.location(mo.row, mo.col)
     elif menu == "inventory":